Use URLs for postcard images (rather than local files)
The inclusion of the postcard jpg within the template skeleton directory was causing RStudio to create a subdirectory for postcards.

I'm thinking that this isn't normally desirable: for example, if you have a GH pages repo that is just a postcard you'd typically want the postcard at the top-level rather than in a subdirectory.

This also makes the Distill case of adding a postcard to a site much more intuitive, since Distill only renders pages at the top level of the repo (so without this the user needs to create the postcard then move it to the site root).

Finally, I think that users will often link to an external image anyway (e.g from their LinkedIn or Twitter profile) so modeling this within the template seems like a good idea.